pyproject.toml
[tool.poetry]
name = "tax-assistant"
version = "0.1.0"
description = ""
authors = ["Dziedzic <michal.dziedzic@joinzoe.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
pydantic = "^2.9.2"
uvicorn = "^0.31.0"
openai = "^1.50.2"
python-dotenv = "^1.0.1"
instructor = "^1.4.3"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
